requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
sentence-transformers
faiss-cpu
huggingface-hub
python-dotenv
